+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `17:37` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `17:36` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `17:36` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `17:35` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`